tkinter-app/requirements.txt
tk
requests
pymongo
flask-socketio
folium
geopy
pywebview
websocket